use a GasCollector to fetch all the data from external parties
