+++ web/src/services/AlertService.ts
@@ -12,7 +12,7 @@ let dbClient: MongoClient | null = null;
 
 export async function Connect(): Promise<MongoClient> {
     if (!dbClient) {
-        dbClient = await MongoClient.connect(AppConfig.MONGODB_CONNECTIONSTRING, { useNewUrlParser: true });
+        dbClient = await MongoClient.connect(AppConfig.MONGODB_CONNECTIONSTRING, { useNewUrlParser: true, useUnifiedTopology: true });
         console.log("alerts Connected..");
     }
